+++ src/sll/stack.c
@@ -149,11 +149,3 @@ void _shift_objects(sll_object_t* o,sll_compilation_data_t* c_dt,sll_object_offs
 	} while (s!=o);
 	*d=*s;
 }
-
-
-
-__SLL_FUNC void sll_setup_stack(sll_stack_data_t* o,sll_char_t* bf,sll_stack_offset_t sz){
-	o->ptr=bf;
-	o->off=0;
-	o->sz=sz;
-}