+++ CHANGELOG.md
@@ -7,6 +7,15 @@
 - Implemented `sll_operator_strict_equal`
 - Strict equal (`(===)`) and not equal (`(!==)`) operators ([#78])
 
+### Changed
+
+- `sll_execute_assembly` now expects the size of the stack instead of the stack itself
+
+### Removed
+
+- Unused stack API: `sll_stack_offset_t`, `sll_stack_data_t`, `sll_deinit_stack_data`, `sll_init_stack_data`, `sll_setup_stack`, `SLL_MAX_STACK_OFFSET`, `SLL_INIT_STACK_DATA` and `SLL_INIT_STACK_DATA_STRUCT`
+- Unused member `sll_stack_offset_t st` from `sll_assembly_instruction_data_t`
+
 ## [0.6.25] - 2021-10-31
 
 ### Added